refactor: Document 컴포넌트 로직 분리
